:recycle: move InMemoryFS.kt back to agents-ext
We don't currently want a dependency from test-utils on rag-base
