+++ agents/agents-ext/src/commonTest/kotlin/ai/koog/agents/ext/tool/file/EditFileToolCoreTest.kt
@@ -2,9 +2,9 @@ package ai.koog.agents.ext.tool.file
 
 import ai.koog.agents.core.tools.DirectToolCallsEnabler
 import ai.koog.agents.core.tools.annotations.InternalAgentToolsApi
+import ai.koog.agents.ext.utils.InMemoryFS
 import ai.koog.rag.base.files.readText
 import ai.koog.rag.base.files.writeText
-import ai.koog.test.utils.InMemoryFS
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertContains
