+++ test-utils/build.gradle.kts
@@ -13,7 +13,6 @@ kotlin {
                 api(kotlin("test"))
                 api(libs.kotlinx.coroutines.test)
                 api(libs.kotlinx.serialization.json)
-                api(project(":rag:rag-base"))
             }
         }
